pyproject.toml
[project]
name = "dish-flow"
version = "0.1.0"
description = ""
authors = [
    {name = "Abboth",email = "Alexmikhailov90@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "redis (>=6.1.0,<7.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pillow (>=11.2.1,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
